[WEB-59] Refactor TagsInputs
TagsInputs previously required another class to extend them to be able
to use the fancy methods and getters they have built-in. This is no
longer valid. TagsInputs now accept all of the same methods that
previously needed to be extended, but as props. For example:

```
new class ExtendedTagsInput extends TagsInput {
  search (query) { ... }

  get valueProp () { ... }
}
```

...can now be written as...

```
<TagsInput
  search={searchFn}
  valueProp="..." />
```